separación de lógica, tabla de proyectos y método delete
